third_party/blink/web_tests/wpt_internal/web-locks/storage-deprecation-trial-enabled.https-expected.txt

This is a testharness.js-based test.
FAIL WebLocks are not partitioned when deprecation trial is enabled assert_equals: expected (boolean) true but got (undefined) undefined
FAIL WebLocks of a nested iframe with a cross-site ancestor are not partitioned when deprecation trial is enabled assert_equals: expected (boolean) true but got (undefined) undefined
Harness: the test ran to completion.
